refactor(dashboard): dashboard 관리 디렉토리 변경
